qr-code-service/requirements.txt
grpcio ~= 1.23
grpcio-tools
qrcode
pillow
